Move sequence errors into src/errors directory
